+++ src/ua.rs
@@ -5,15 +5,19 @@ mod client;
 mod data_types;
 mod monitored_item_id;
 mod node_class_mask;
+mod secure_channel_state;
+mod session_state;
 mod subscription_id;
 mod values;
 
 pub use self::{
     array::Array,
-    client::Client,
+    client::{Client, ClientState},
     data_types::*,
     monitored_item_id::MonitoredItemId,
     node_class_mask::NodeClassMask,
+    secure_channel_state::SecureChannelState,
+    session_state::SessionState,
     subscription_id::SubscriptionId,
     values::{ArrayValue, ScalarValue, VariantValue},
 };